Refactor user_for_paper_trail method and history view layout
This commit refactors the `user_for_paper_trail` method in the `ApplicationController` to simplify the code. It also updates the `history.arb` view layout to improve readability and remove unnecessary code. Additionally, a new file `paper_trail.rb` is added to the `config/initializers` directory to disable `PaperTrail` when running in the Rails console.